stabilizing LM algorithm, added uncertainty calculation

+++ autoprof/image/image_object.py
@@ -141,7 +141,7 @@ class BaseImage(object):
         hdul = fits.open(filename)
         for hdu in hdul:
             if "IMAGE" in hdu.header and hdu.header["IMAGE"] == "PRIMARY":
-                self.set_data(np.array(hdu.data, dtype = np.float64), require_shape = False)
+                self.set_data(np.array(hdu.data, dtype = np.float32), require_shape = False)
                 self.pixelscale = eval(hdu.header.get("PXLSCALE"))
                 self.zeropoint = eval(hdu.header.get("ZEROPNT"))
                 self.note = hdu.header.get("NOTE")
